requirements.txt
Flask==3.0.3
Pillow==10.3.0
PyMuPDF==1.24.4
Werkzeug==3.0.3
